ai-agent/requirements.txt
librosa==0.10.1
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2
soundfile==0.12.1
pinecone-client==2.2.4
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
httpx==0.25.1
pytest==7.4.3
pytest-cov==4.1.0
prophet==1.1.5
pandas==2.1.3
web3==6.11.3
eth-account==0.10.0
